Add path information to NetCDFInputData (#114)
* Create Path column only in standalone plotter

* Implement recursive var search in netcdf

* Refactor to use variable wrappers

* Refactor into multiple files
