+++ Src/GUI/Plugins/PlotterPlugin.py
@@ -34,6 +34,7 @@ from MDANSE.GUI.Plugins.Plotter1D import Plotter1D
 from MDANSE.GUI.Plugins.Plotter2D import Plotter2D
 from MDANSE.GUI.Plugins.Plotter3D import Plotter3D
 from MDANSE.GUI.Icons import ICONS
+from MDANSE.IO.IOUtils import load_variables
 
 
 class PlotterError(Error):
@@ -92,7 +93,8 @@ class MultiViewPanel(wx.Panel):
 
     def Update(self):
         self._mgr.Update()
-        
+
+
 class DataPanel(wx.Panel):
 
     def __init__(self, parent, *args,**kwds):
@@ -478,24 +480,7 @@ class PlotterFrame(wx.Frame):
 
             ext = os.path.splitext(filename)[1]
             with PLOTTER_DATA_TYPES[ext](filename, 'r') as f:
-                data = collections.OrderedDict()
-                for vname, vinfo in f.variables.items():
-                    vpath, variable = vinfo
-                    arr = variable.get_array()
-                    attributes = variable.get_attributes()
-
-                    data[vname] = {}
-                    if 'axis' in attributes:
-                        axis = attributes['axis']
-                        if axis:
-                            data[vname]['axis'] = axis.split('|')
-                        else:
-                            data[vname]['axis'] = []
-                    else:
-                        data[vname]['axis'] = []
-                    data[vname]['path'] = vpath
-                    data[vname]['data'] = arr
-                    data[vname]['units'] = attributes.get('units', 'au')
+                data = load_variables(f.variables)
 
                 unique_name = self.unique(basename, self.plugin._dataDict)
